lesson-1-2 assigment-components
