+++ src/App.js
@@ -1,22 +1,17 @@
 import React from "react";
+import AddTodoForm from "./AddTodoForm";
+import TodoList from "./ToDoList";
 
-const todoList = [];
 
-todoList[0] = { id: "1", title: "Complete assignment" };
-todoList[1] = { id: "2", title: "Upload to Github" };
-todoList[2] = { id: "3", title: "Upload the link" };
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map((item) => {
-          return <li key={item.id.toString()}> {item.title}</li>;
-        })}
-      </ul>
+        <AddTodoForm/>
+        <TodoList />
     </div>
-  );
+  ); 
 }
 
 export default App;
